execution-service/requirements.txt
pandas
sqlalchemy
psycopg2-binary
pika